[MOD/#378] 타겟아이디 서버 명세 변경 대응

+++ presentation/notification/src/main/java/com/hilingual/presentation/notification/main/NotificationScreen.kt
@@ -61,8 +61,8 @@ internal fun NotificationRoute(
         onFeedNotificationClick = { notification ->
             viewModel.readNotification(notification.id)
             when (notification.feedType) {
-                FeedNotificationType.LIKE_DIARY -> navigateToFeedDiary(notification.targetId.toLong())
-                FeedNotificationType.FOLLOW_USER -> navigateToFeedProfile(notification.targetId.toLong())
+                FeedNotificationType.LIKE_DIARY -> navigateToFeedDiary(notification.targetId)
+                FeedNotificationType.FOLLOW_USER -> navigateToFeedProfile(notification.targetId)
             }
         },
         onNoticeNotificationClick = { notification -> navigateToNoticeDetail(notification.id) },
